UX: Make the event text colors more readable
It's not ideal, but at least should handle worst cases.
